Attempt to stabilize scheduler job registration

+++ thefederation/apps.py
@@ -1,4 +1,5 @@
 import datetime
+import sys
 
 import django_rq
 from django.apps import AppConfig
@@ -9,11 +10,15 @@ class TheFederationConfig(AppConfig):
     verbose_name = "The Federation"
 
     def ready(self):
+        # Only register tasks if RQ Scheduler process
+        if "rqscheduler" not in sys.argv:
+            return
+
         from thefederation.social import make_daily_post
         from thefederation.tasks import aggregate_daily_stats
         from thefederation.tasks import poll_nodes
 
-        scheduler = django_rq.get_scheduler('high')
+        scheduler = django_rq.get_scheduler()
         # Delete any existing jobs in the scheduler when the app starts up
         for job in scheduler.get_jobs():
             job.delete()
@@ -22,6 +27,7 @@ class TheFederationConfig(AppConfig):
             scheduled_time=datetime.datetime.utcnow(),
             func=aggregate_daily_stats,
             interval=5500,
+            queue_name='high',
         )
         scheduler.cron(
             '0 10 * * *',
@@ -29,10 +35,9 @@ class TheFederationConfig(AppConfig):
             queue_name='high',
         )
 
-        scheduler = django_rq.get_scheduler('medium')
-
         scheduler.schedule(
             scheduled_time=datetime.datetime.utcnow(),
             func=poll_nodes,
             interval=10800,
+            queue_name='medium',
         )
